pyproject.toml
[project]
name = "sentinelx"
version = "1.0.0"
description = "Financial Fraud Investigation Environment with Adaptive Adversary — OpenEnv hackathon submission"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Abhay Maheshwari", email = "maheshwariabhay9@gmail.com" }
]
keywords = ["openenv", "reinforcement-learning", "fraud-detection", "llm", "agent"]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentinelx"]

[project.scripts]
sentinelx-server = "sentinelx.server.app:app"
